Extract ImportantNotesBar to its own file and make copy changes
